Change role IDs to i32
Changed `admin_role_id` and `student_role_id` from i16 to i32 in entity definitions and migrations. Adjusted `create_default_admin` method to accept email and password parameters.

+++ entity/src/admins.rs
@@ -13,7 +13,7 @@ pub struct Model {
     #[sea_orm(unique)]
     pub email: String,
     pub password_hash: String,
-    pub admin_role_id: i16,
+    pub admin_role_id: i32,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
